Adjust styles for kramdown compatibility
The developer guide was originally styled with
GFM syntax, which rendered perfectly in GFM.
However, GitHub Pages deployments utilize
kramdown, leading to discrepancies in the
rendering.

Let's modify the Table of Contents and
introduction box styles to ensure compatibility
with kramdown. This change will make certain
that our guide is displayed consistently and
correctly on GitHub Pages.

Adapting to kramdown ensures the documentation
remains professional and accessible to all
users, irrespective of the markdown processor
used in the backend.